fix list
