+++ ft_list/ft_lstdelone.c
@@ -6,11 +6,11 @@
 /*   By: reasuke <reasuke@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/04 19:30:04 by reasuke           #+#    #+#             */
-/*   Updated: 2023/09/06 12:29:52 by reasuke          ###   ########.fr       */
+/*   Updated: 2024/02/24 12:16:48 by reasuke          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include "ft_list.h"
 
 void	ft_lstdelone(t_list *lst, void (*f)(void *))
 {
