+++ ft_list/ft_lstiter.c
@@ -1,41 +1,24 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_lstmap.c                                        :+:      :+:    :+:   */
+/*   ft_lstiter.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: reasuke <reasuke@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/09/04 22:19:08 by reasuke           #+#    #+#             */
-/*   Updated: 2023/09/29 14:24:38 by reasuke          ###   ########.fr       */
+/*   Created: 2023/09/04 19:12:35 by reasuke           #+#    #+#             */
+/*   Updated: 2024/02/24 12:16:53 by reasuke          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include "ft_list.h"
 
-t_list	*ft_lstmap(t_list *lst, void *(*f)(void *), void (*del)(void *))
+void	ft_lstiter(t_list *lst, void (*f)(void *))
 {
-	t_list	*ret;
-	t_list	*new_lst;
-	t_list	*cur;
-
-	if (!f || !del)
-		return (NULL);
-	ret = NULL;
-	cur = NULL;
+	if (!f)
+		return ;
 	while (lst)
 	{
-		new_lst = ft_lstnew(f(lst->content));
-		if (!new_lst)
-		{
-			ft_lstclear(&new_lst, del);
-			return (NULL);
-		}
-		if (cur)
-			cur->next = new_lst;
-		else
-			ret = new_lst;
-		cur = new_lst;
+		f(lst->content);
 		lst = lst->next;
 	}
-	return (ret);
 }
