+++ ft_list/ft_lstlast.c
@@ -6,11 +6,11 @@
 /*   By: reasuke <reasuke@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/04 14:19:25 by reasuke           #+#    #+#             */
-/*   Updated: 2023/09/04 19:21:11 by reasuke          ###   ########.fr       */
+/*   Updated: 2024/02/24 12:16:57 by reasuke          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include "ft_list.h"
 
 t_list	*ft_lstlast(t_list *lst)
 {
