+++ ft_list/ft_lstnew.c
@@ -1,26 +1,25 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_lstadd_front.c                                  :+:      :+:    :+:   */
+/*   ft_lstnew.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: reasuke <reasuke@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/09/04 19:06:44 by reasuke           #+#    #+#             */
-/*   Updated: 2023/11/09 15:34:33 by reasuke          ###   ########.fr       */
+/*   Created: 2023/09/04 13:08:55 by reasuke           #+#    #+#             */
+/*   Updated: 2024/02/24 12:17:06 by reasuke          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include "ft_list.h"
 
-void	ft_lstadd_front(t_list **lst, t_list *new_lst)
+t_list	*ft_lstnew(void *content)
 {
-	if (!lst || !new_lst)
-		return ;
-	if (!*lst)
-	{
-		*lst = new_lst;
-		return ;
-	}
-	new_lst->next = *lst;
-	*lst = new_lst;
+	t_list	*ret;
+
+	ret = malloc(sizeof(t_list));
+	if (!ret)
+		return (NULL);
+	ret->content = content;
+	ret->next = NULL;
+	return (ret);
 }
