+++ test/unit/test_lst_before.cpp
@@ -4,7 +4,7 @@
 #include "unit_test_utils.hpp"
 
 extern "C" {
-#include "libft.h"
+#include "ft_list.h"
 }
 
 static void *convert_const_char_to_void(const char *str) {
